requirements.txt
fastapi
uvicorn[standard]
passlib
python-dotenv
python-jose[cryptography]
python-multipart